fix: Correction Request Delete

+++ AppDiv.CRVS.Application/Features/CorrectionRequests/Commands/Delete/DeleteCorrectionRequestCommad.cs
@@ -1,7 +1,9 @@
 using AppDiv.CRVS.Application.Common;
+using AppDiv.CRVS.Application.Exceptions;
 using AppDiv.CRVS.Application.Interfaces.Persistence;
 using AppDiv.CRVS.Domain.Repositories;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -33,13 +35,15 @@ namespace AppDiv.CRVS.Application.Features.CorrectionRequests.Commands.Delete
             var response = new BaseResponse();
             try
             {   foreach(Guid id in request.Ids){
-                var correctionRequest=await _correctionRequestRepository.GetAsync(id);
+                var correctionRequest= _correctionRequestRepository.GetAll()
+                .Include(x=>x.Request).Where(x=>x.Id==id).FirstOrDefault();
+                if(correctionRequest.Request.currentStep!=0){   
+                    throw  new NotFoundException("You Can not delete this Request It Is Approved");
+                }
                 await _requestRepostory.DeleteAsync(correctionRequest.RequestId);
                 await _correctionRequestRepository.DeleteAsync(id);
-            }
-                
+            }  
                 await _correctionRequestRepository.SaveChangesAsync(cancellationToken);
-
                 response.Deleted("Correctoon Request");
 
             }
